CMF

Background

The Cluster Mean Field (CMF) approach used in this work was developed by Jimènez-Hoyos and Scuseria.

Algorithm

Index

FermiCG.cmf_ci

FermiCG.cmf_ci

FermiCG.cmf_ci

FermiCG.cmf_ci

FermiCG.cmf_ci_iteration

FermiCG.cmf_ci_iteration

FermiCG.cmf_oo

FermiCG.cmf_oo

FermiCG.cmf_oo

FermiCG.compute_cmf_energy

FermiCG.compute_cmf_energy

FermiCG.form_1rdm_dressed_ints

FermiCG.form_casci_ints

Documentation

FermiCG.cmf_ci — Method

cmf_ci(ints, clusters, fspace, dguess; 
        max_iter=10, dconv=1e-6, econv=1e-10, verbose=1)

Optimize the 1RDM for CMF-CI

FermiCG.cmf_ci — Method

cmf_ci(ints, clusters, fspace, in_rdm1a, in_rdm1b; 
            max_iter=10, dconv=1e-6, econv=1e-10, verbose=1,sequential=false)

Optimize the 1RDM for CMF-CI

#Arguments

ints::InCoreInts: integrals for full system

clusters::Vector{Cluster}: vector of cluster objects

fspace::Vector{Vector{Integer}}: vector of particle number occupations for each cluster specifying the sectors of fock space

in_rdm1a: initial guess for 1particle density matrix for alpha electrons

in_rdm1b: initial guess for 1particle density matrix for beta electrons

dconv: Convergence threshold for change in density

econv: Convergence threshold for change in energy

sequential: Use the density matrix of the previous cluster in a cMF iteration to form effective integrals. Improves comvergence, may depend on cluster orderings

verbose: Printing level

FermiCG.cmf_ci — Method

cmf_ci(mol::Molecule, C::Matrix, clusters::Vector{Cluster}, fspace::Vector, dguess; 
        max_iter=10, dconv=1e-6, econv=1e-10, verbose=1)

Optimize the 1RDM for CMF-CI, without requiring an InCoreInts object

Arguments

mol::Molecule: a FermiCG.Molecule type

C: MO coefficients for full system (spin restricted)

clusters::Vector{Cluster}: vector of Cluster objects

fspace::Vector{Vector{Integer}}: vector of particle number occupations for each cluster specifying the sectors of fock space

dguess: initial guess for 1particle density matrix (spin summed)

dconv: Convergence threshold for change in density

econv: Convergence threshold for change in energy

verbose: how much to print

FermiCG.cmf_ci — Method

cmf_ci(mol::Molecule, C::Matrix, clusters::Vector{Cluster}, fspace::Vector, dguess; 
        max_iter=10, dconv=1e-6, econv=1e-10, verbose=1)

Optimize the 1RDM for CMF-CI, without requiring an InCoreInts object

Arguments

mol::Molecule: a FermiCG.Molecule type

C: MO coefficients for full system (spin restricted)

clusters::Vector{Cluster}: vector of Cluster objects

fspace::Vector{Vector{Integer}}: vector of particle number occupations for each cluster specifying the sectors of fock space

dguess: initial guess for 1particle density matrix (spin summed)

dconv: Convergence threshold for change in density

econv: Convergence threshold for change in energy

verbose: how much to print

FermiCG.cmf_ci_iteration — Method

cmf_ci_iteration(ints::InCoreInts, clusters::Vector{Cluster}, rdm1a, rdm1b, fspace; verbose=1)

Perform single CMF-CI iteration, returning new energy, and density

FermiCG.cmf_ci_iteration — Method

cmf_ci_iteration(mol::Molecule, C, rdm1a, rdm1b, clusters, fspace; 
                 verbose=1)

Perform single CMF-CI iteration, returning new energy, and density. This method forms the eri's on the fly to avoid global N^4 storage

Arguments

mol::Molecule: a FermiCG.Molecule type

C: MO coefficients for full system (spin restricted)

rdm1a: 1particle density matrix (alpha)

rdm1b: 1particle density matrix (beta)

clusters::Vector{Cluster}: vector of Cluster objects

fspace::Vector{Vector{Int}}: vector of particle number occupations for each cluster specifying the sectors of fock space

verbose: how much to print

See also: cmf_ci_iteration

FermiCG.cmf_oo — Method

cmf_oo(ints::InCoreInts, clusters::Vector{Cluster}, fspace, dguess_a, dguess_b; 
            max_iter_oo=100, max_iter_ci=100, gconv=1e-6, verbose=0, method="bfgs", alpha=nothing,sequential=false)

Do CMF with orbital optimization

#Arguments

ints::InCoreInts: integrals for full system

clusters::Vector{Cluster}: vector of cluster objects

fspace::Vector{Vector{Integer}}: vector of particle number occupations for each cluster specifying the sectors of fock space

dguess_a: initial guess for 1particle density matrix for alpha electrons

dguess_b: initial guess for 1particle density matrix for beta electrons

max_iter_oo: Max iter for the orbital optimization iterations

max_iter_ci: Max iter for the cmf iteration for the cluster states

gconv: Convergence threshold for change in gradient of energy

sequential: If true use the density matrix of the previous cluster in a cMF iteration to form effective integrals. Improves comvergence, may depend on cluster orderings

verbose: Printing level

method: optimization method

FermiCG.cmf_oo — Method

cmf_oo(ints::InCoreInts, clusters::Vector{Cluster}, fspace, dguess; 
        max_iter_oo=100, max_iter_ci=100, gconv=1e-6, verbose=0, method="bfgs")

Do CMF with orbital optimization

FermiCG.cmf_oo — Method

cmf_oo(mol::Molecule, Cguess::Matrix, clusters::Vector{Cluster}, fspace, dguess; 
        max_iter_oo=100, max_iter_ci=100, gconv=1e-6, verbose=0, method="bfgs")

Do CMF with orbital optimization with on the fly integrals

FermiCG.compute_cmf_energy — Method

compute_cmf_energy(ints, rdm1s, rdm2s, clusters)

Compute the energy of a cluster-wise product state (CMF), specified by a list of 1 and 2 particle rdms local to each cluster. This method uses the full system integrals.

ints::InCoreInts: integrals for full system

rdm1s: dictionary (ci.idx => Array) of 1rdms from each cluster (spin summed)

rdm2s: dictionary (ci.idx => Array) of 2rdms from each cluster (spin summed)

clusters::Vector{Cluster}: vector of cluster objects

return the total CMF energy

FermiCG.compute_cmf_energy — Method

compute_cmf_energy(mol::Molecule, C::Matrix, rdm1s, rdm2s, clusters)

Compute the energy of a cluster-wise product state (CMF), specified by a list of 1 and 2 particle rdms local to each cluster

#Arguments

mol::Molecule

C::Matrix: MO coefficients

rdm1s: dictionary (ci.idx => Array) of 1rdms from each cluster (spin summed)

rdm2s: dictionary (ci.idx => Array) of 2rdms from each cluster (spin summed)

clusters::Vector{Cluster}: vector of cluster objects

return the total CMF energy

FermiCG.form_1rdm_dressed_ints — Method

form_1rdm_dressed_ints(ints::InCoreInts, orb_list, rdm1a, rdm1b)

Obtain a subset of integrals which act on the orbitals in Cluster, embedding the 1rdm from the rest of the system

Returns an InCoreInts type

FermiCG.form_casci_ints — Method

form_casci_ints(ints::InCoreInts, ci::Cluster, rdm1a, rdm1b)

Obtain a subset of integrals which act on the orbitals in Cluster, embedding the 1rdm from the rest of the system

Returns an InCoreInts type