FermiCG.ClusteredOperator

FermiCG.ClusteredTerm

FermiCG.extract_1body_operator

FermiCG.extract_ClusteredTerms

FermiCG.extract_S2

FermiCG.unique!

ClusteredTerms

FermiCG.ClusteredOperator — Type

trans::Dict{TransferConfig,Vector{ClusteredTerm}}

FermiCG.ClusteredTerm — Type

ops::Tuple{String}
delta::TransferConfig{1}
parity::Tuple{Int}
clusters::Tuple{Cluster}
ints::Array{Float64}
cache::Dict

input:

delta = list of change of Na,Nb,state

e.g., [(-1,-1),(1,1),(0,0)] means alpha and beta transition
		from cluster 1 to 2, cluster 3 is fock diagonal

ops = list of operators

e.g., ["ab","AB",""]

ints = tensor containing the integrals for this block

e.g., ndarray([p,q,r,s]) where p,q are in 1 and r,s are in 2

data contained in object

active: list of clusters which have non-identity operators
		this includes fock-diagonal couplings,
		e.g., ["Aa","","Bb"] would have active = [0,2]

parity: does each operator have even or odd number of second quantized operators

FermiCG.extract_1body_operator — Method

extract_1body_operator(clustered_ham::ClusteredOperator; op_string="H")

Extract a 1-body operator for use in perturbation theory

op_string: either H or Hcmf

FermiCG.extract_ClusteredTerms — Method

extract_terms(ints::InCoreInts, clusters)

Extract all ClusteredTerm types from a given 1e integral tensor and a list of clusters returns terms::Dict{TransferConfig,Vector{ClusteredTerm}}

FermiCG.extract_S2 — Method

extract_S2(clusters)

Form a clustered operator type for the S^2 operator

FermiCG.unique! — Method

unique!(clustered_ham::ClusteredOperator)

combine terms to keep only unique operators