FermiCG.Cluster

FermiCG.ClusterBasis

FermiCG.ClusterSubspace

Base.Multimedia.display

Base.length

FermiCG.compute_cluster_eigenbasis

FermiCG.compute_cluster_ops

FermiCG.dim_tot

FermiCG.dim_tot

FermiCG.get_ortho_compliment

FermiCG.possible_focksectors

FermiCG.rotate!

FermiCG.rotate!

FermiCG.tdm_A

FermiCG.tdm_AA

FermiCG.tdm_AAa

FermiCG.tdm_AB

FermiCG.tdm_ABa

FermiCG.tdm_Aa

FermiCG.tdm_Ab

FermiCG.tdm_H

FermiCG.tdm_S2

Clusters

FermiCG.Cluster — Type

idx::Int
orb_list::Vector{Int}

FermiCG.ClusterBasis — Type

cluster::Cluster                            # Cluster to which basis belongs
basis::Dict{Tuple,Matrix{Float64}}          # Basis vectors (nα, nβ)=>[I,s]

These basis coefficients map local slater determinants to local vectors (nα, nβ): V[αstring*βstring, cluster_state]

FermiCG.ClusterSubspace — Type

Defines a single cluster's subspace for Tucker. Each focksector is allowed to have a distinct cluster state range for the subspace.

cluster::Cluster
data::OrderedDict{Tuple{UInt8,UInt8}, UnitRange{Int}}

Base.Multimedia.display — Method

display(c::Cluster)

Base.length — Method

length(c::Cluster)

Return number of orbitals in Cluster

FermiCG.compute_cluster_eigenbasis — Method

compute_cluster_eigenbasis(ints::InCoreInts, clusters::Vector{Cluster}; 
    init_fspace=nothing, delta_elec=nothing, verbose=0, max_roots=10, 
    rdm1a=nothing, rdm1b=nothing)

Return a Vector of ClusterBasis for each Cluster

ints::InCoreInts: In-core integrals

clusters::Vector{Cluster}: Clusters

verbose::Int: Print level

init_fspace: list of pairs of (nα,nβ) for each cluster for defining reference space for selecting out only certain fock sectors

delta_elec: number of electrons different from reference (init_fspace)

max_roots::Int: Maximum number of vectors for each focksector basis

rdm1a: background density matrix for embedding local hamiltonian (alpha)

rdm1b: background density matrix for embedding local hamiltonian (beta)

FermiCG.compute_cluster_ops — Method

compute_cluster_ops(cluster_bases::Vector{ClusterBasis})

FermiCG.dim_tot — Method

dim_tot(c::Cluster, na, nb)

Return dimension of hilbert space spanned by number of orbitals in Cluster with na and nb number of alpha/beta electrons.

FermiCG.dim_tot — Method

dim_tot(c::Cluster)

Return dimension of hilbert space spanned by number of orbitals in Cluster. This is all sectors

FermiCG.get_ortho_compliment — Method

get_ortho_compliment(tss::ClusterSubspace, cb::ClusterBasis)

For a given ClusterSubspace, tss, return the subspace remaining

FermiCG.possible_focksectors — Method

possible_focksectors(c::Cluster, delta_elec=nothing)

Get list of possible fock spaces accessible to the cluster

delta_elec::Vector{Int}: (nα, nβ, Δ) restricts fock spaces to: (nα,nβ) ± Δ electron transitions

FermiCG.rotate! — Method

rotate!(cb::ClusterBasis, U::Dict{Tuple,Matrix{T}}) where {T}

Rotate cb by unitary matrices in U

FermiCG.rotate! — Method

rotate!(ops::ClusterOps, U::Dict{Tuple,Matrix{T}}) where {T}

Rotate ops by unitary matrices in U

FermiCG.tdm_A — Method

tdm_A(cb::ClusterBasis; verbose=0)

Compute <s|p'|t> between all cluster states, s and t from accessible sectors of a cluster's fock space.

Returns Dict[((na,nb),(na,nb))] => Array

FermiCG.tdm_AA — Method

tdm_AA(cb::ClusterBasis; verbose=0)

Compute <s|p'q'|t> between all cluster states, s and t from accessible sectors of a cluster's fock space.

Returns Dict[((na,nb),(na,nb))] => Array

FermiCG.tdm_AAa — Method

tdm_AAa(cb::ClusterBasis, spin_case; verbose=0)

Compute <s|p'q'r|t> between all cluster states, s and t from accessible sectors of a cluster's fock space.

spin_case: alpha or beta

Returns Dict[((na,nb),(na,nb))] => Array

FermiCG.tdm_AB — Method

tdm_AB(cb::ClusterBasis; verbose=0)

Compute <s|p'q'|t> between all cluster states, s and t from accessible sectors of a cluster's fock space, where p' is alpha and q' is beta.

Returns Dict[((na,nb),(na,nb))] => Array

FermiCG.tdm_ABa — Method

tdm_ABa(cb::ClusterBasis, spin_case; verbose=0)

Compute <s|p'q'r|t> between all cluster states, s and t from accessible sectors of a cluster's fock space.

spin_case: alpha or beta

Returns Dict[((na,nb),(na,nb))] => Array

FermiCG.tdm_Aa — Method

tdm_Aa(cb::ClusterBasis, spin_case; verbose=0)

Compute <s|p'q|t> between all cluster states, s and t from accessible sectors of a cluster's fock space.

spin_case: alpha or beta

Returns Dict[((na,nb),(na,nb))] => Array

FermiCG.tdm_Ab — Method

tdm_Ab(cb::ClusterBasis; verbose=0)

Compute <s|p'q|t> between all cluster states, s and t from accessible sectors of a cluster's fock space, where p' is alpha and q is beta.

Returns Dict[((na,nb),(na,nb))] => Array

FermiCG.tdm_H — Method

tdm_H(cb::ClusterBasis; verbose=0)

Compute local Hamiltonian <s|H|t> between all cluster states, s and t from accessible sectors of a cluster's fock space.

Returns Dict[((na,nb),(na,nb))] => Array

FermiCG.tdm_S2 — Method