FermiCG.Atom

FermiCG.InCoreInts

FermiCG.Molecule

FermiCG.compute_energy

FermiCG.compute_energy

FermiCG.orbital_rotation

FermiCG.orbital_rotation!

FermiCG.subset

Hamiltonians

FermiCG.Atom — Type

id::Integer             #index of atom in the molecule
symbol::String          #Atomic ID (E.g. H, He, ...)
xyz::Array{Float64,1}   #list of XYZ coordinates

Simply an Atom

FermiCG.InCoreInts — Type

h0::Real                # constant energy shift
h1::Array{Float64,2}    # one electron integrals
h2::Array{Float64,4}    # two electron integrals (chemist's notation)

Type to hold a second quantized Hamiltonian coefficients in memory

FermiCG.Molecule — Type

charge::Integer         #overall charge on molecule
multiplicity::Integer   #2S+1
atoms::Vector{Atom}     #Vector of `Atoms`
basis::String           #Basis set

Molecule essentially as a Vector of atoms, number of electrons and basis set

FermiCG.compute_energy — Method

compute_energy(h0, h1, h2, rdm1, rdm2)

Given an energy shift h0, 1e integrals h1, and 2e ints h2 along with a 1rdm and 2rdm on the same space, return the energy

FermiCG.compute_energy — Method

compute_energy(ints::InCoreInts, rdm1, rdm2)

Return energy defined by rdm1 and rdm2

FermiCG.orbital_rotation! — Method

orbital_rotation!(ints::InCoreInts, U)

Transform electronic integrals, by U i.e.,

\[h_{pq} = U_{rp}h_{rs}U_{sq}\]

\[(pq|rs) = (tu|vw)U_{tp}U_{uq}U_{vr}U_{ws}\]

FermiCG.orbital_rotation — Method

orbital_rotation(ints::InCoreInts, U)

Transform electronic integrals, by U i.e.,

\[h_{pq} = U_{rp}h_{rs}U_{sq}\]

\[(pq|rs) = (tu|vw)U_{tp}U_{uq}U_{vr}U_{ws}\]

FermiCG.subset — Function

subset(ints::InCoreInts, list; rmd1a=nothing, rdm1b=nothing)

Extract a subset of integrals acting on orbitals in list, returned as InCoreInts type

ints::InCoreInts: Integrals for full system

list: list of orbital indices in subset

rdm1a: 1RDM for embedding α density to make CASCI hamiltonian

rdm1b: 1RDM for embedding β density to make CASCI hamiltonian