Internals

FermiCG.Atom

FermiCG.Cluster

FermiCG.ClusterBasis

FermiCG.ClusterSubspace

FermiCG.InCoreInts

FermiCG.Molecule

Base.Multimedia.display

Base.length

FermiCG.compute_cluster_eigenbasis

FermiCG.compute_cluster_est_basis

FermiCG.compute_cluster_ops

FermiCG.compute_energy

FermiCG.compute_energy

FermiCG.dim_tot

FermiCG.dim_tot

FermiCG.form_schmidt_basis

FermiCG.get_ortho_compliment

FermiCG.orbital_rotation

FermiCG.orbital_rotation!

FermiCG.possible_focksectors

FermiCG.rotate!

FermiCG.rotate!

FermiCG.subset

FermiCG.tdm_A

FermiCG.tdm_AA

FermiCG.tdm_AAa

FermiCG.tdm_AB

FermiCG.tdm_ABa

FermiCG.tdm_Aa

FermiCG.tdm_Ab

FermiCG.tdm_H

FermiCG.tdm_S2

Clusters

Clusters are simply collections of orbitals.

FermiCG.Cluster — Type

idx::Int
orb_list::Vector{Int}

FermiCG.ClusterBasis — Type

cluster::Cluster                            # Cluster to which basis belongs
basis::Dict{Tuple,Matrix{Float64}}          # Basis vectors (nα, nβ)=>[I,s]

These basis coefficients map local slater determinants to local vectors (nα, nβ): V[αstring*βstring, cluster_state]

FermiCG.ClusterSubspace — Type

Defines a single cluster's subspace for Tucker. Each focksector is allowed to have a distinct cluster state range for the subspace.

cluster::Cluster
data::OrderedDict{Tuple{UInt8,UInt8}, UnitRange{Int}}

Base.Multimedia.display — Method

display(c::Cluster)

Base.length — Method

length(c::Cluster)

Return number of orbitals in Cluster

FermiCG.compute_cluster_eigenbasis — Method

compute_cluster_eigenbasis(ints::InCoreInts, clusters::Vector{Cluster}; 
    init_fspace=nothing, delta_elec=nothing, verbose=0, max_roots=10, 
    rdm1a=nothing, rdm1b=nothing)

Return a Vector of ClusterBasis for each Cluster

ints::InCoreInts: In-core integrals

clusters::Vector{Cluster}: Clusters

verbose::Int: Print level

init_fspace: list of pairs of (nα,nβ) for each cluster for defining reference space for selecting out only certain fock sectors

delta_elec: number of electrons different from reference (init_fspace)

max_roots::Int: Maximum number of vectors for each focksector basis

rdm1a: background density matrix for embedding local hamiltonian (alpha)

rdm1b: background density matrix for embedding local hamiltonian (beta)

FermiCG.compute_cluster_est_basis — Method

compute_cluster_est_basis(ints::InCoreInts, clusters::Vector{Cluster}; 
    init_fspace=nothing, delta_elec=nothing, verbose=0, max_roots=10, 
    rdm1a=nothing, rdm1b=nothing)

Return a Vector of ClusterBasis for each Cluster using the Embedded Schmidt Truncation

ints::InCoreInts: In-core integrals

clusters::Vector{Cluster}: Clusters

Da: background density matrix for embedding local hamiltonian (alpha)

Db: background density matrix for embedding local hamiltonian (beta)

init_fspace: list of pairs of (nα,nβ) for each cluster for defining reference space for selecting out only certain fock sectors

thresh_schmidt: the threshold for the EST

thresh_orb: threshold for the orbital

thresh_ci: threshold for the ci problem

FermiCG.compute_cluster_ops — Method

compute_cluster_ops(cluster_bases::Vector{ClusterBasis})

FermiCG.dim_tot — Method

dim_tot(c::Cluster, na, nb)

Return dimension of hilbert space spanned by number of orbitals in Cluster with na and nb number of alpha/beta electrons.

FermiCG.dim_tot — Method

dim_tot(c::Cluster)

Return dimension of hilbert space spanned by number of orbitals in Cluster. This is all sectors

FermiCG.form_schmidt_basis — Method

form_schmidt_basis

threshorb : threshold for determining how many bath orbitals to include threshschmidt : threshold for determining how many singular vectors to include for cluster basis

Returns new basis for the cluster

FermiCG.get_ortho_compliment — Method

get_ortho_compliment(tss::ClusterSubspace, cb::ClusterBasis)

For a given ClusterSubspace, tss, return the subspace remaining

FermiCG.possible_focksectors — Method

possible_focksectors(c::Cluster, delta_elec=nothing)

Get list of possible fock spaces accessible to the cluster

delta_elec::Vector{Int}: (nα, nβ, Δ) restricts fock spaces to: (nα,nβ) ± Δ electron transitions

FermiCG.rotate! — Method

rotate!(cb::ClusterBasis, U::Dict{Tuple,Matrix{T}}) where {T}

Rotate cb by unitary matrices in U

FermiCG.rotate! — Method

rotate!(ops::ClusterOps, U::Dict{Tuple,Matrix{T}}) where {T}

Rotate ops by unitary matrices in U

FermiCG.tdm_A — Method

tdm_A(cb::ClusterBasis; verbose=0)

Compute <s|p'|t> between all cluster states, s and t from accessible sectors of a cluster's fock space.

Returns Dict[((na,nb),(na,nb))] => Array

FermiCG.tdm_AA — Method

tdm_AA(cb::ClusterBasis; verbose=0)

Compute <s|p'q'|t> between all cluster states, s and t from accessible sectors of a cluster's fock space.

Returns Dict[((na,nb),(na,nb))] => Array

FermiCG.tdm_AAa — Method

tdm_AAa(cb::ClusterBasis, spin_case; verbose=0)

Compute <s|p'q'r|t> between all cluster states, s and t from accessible sectors of a cluster's fock space.

spin_case: alpha or beta

Returns Dict[((na,nb),(na,nb))] => Array

FermiCG.tdm_AB — Method

tdm_AB(cb::ClusterBasis; verbose=0)

Compute <s|p'q'|t> between all cluster states, s and t from accessible sectors of a cluster's fock space, where p' is alpha and q' is beta.

Returns Dict[((na,nb),(na,nb))] => Array

FermiCG.tdm_ABa — Method

tdm_ABa(cb::ClusterBasis, spin_case; verbose=0)

Compute <s|p'q'r|t> between all cluster states, s and t from accessible sectors of a cluster's fock space.

spin_case: alpha or beta

Returns Dict[((na,nb),(na,nb))] => Array

FermiCG.tdm_Aa — Method

tdm_Aa(cb::ClusterBasis, spin_case; verbose=0)

Compute <s|p'q|t> between all cluster states, s and t from accessible sectors of a cluster's fock space.

spin_case: alpha or beta

Returns Dict[((na,nb),(na,nb))] => Array

FermiCG.tdm_Ab — Method

tdm_Ab(cb::ClusterBasis; verbose=0)

Compute <s|p'q|t> between all cluster states, s and t from accessible sectors of a cluster's fock space, where p' is alpha and q is beta.

Returns Dict[((na,nb),(na,nb))] => Array

FermiCG.tdm_H — Method

tdm_H(cb::ClusterBasis; verbose=0)

Compute local Hamiltonian <s|H|t> between all cluster states, s and t from accessible sectors of a cluster's fock space.

Returns Dict[((na,nb),(na,nb))] => Array

FermiCG.tdm_S2 — Method

Hamiltonians

These are types/methods to handle the fermionic Hamiltonian

FermiCG.Atom — Type

id::Integer             #index of atom in the molecule
symbol::String          #Atomic ID (E.g. H, He, ...)
xyz::Array{Float64,1}   #list of XYZ coordinates

Simply an Atom

FermiCG.InCoreInts — Type

h0::Real                # constant energy shift
h1::Array{Float64,2}    # one electron integrals
h2::Array{Float64,4}    # two electron integrals (chemist's notation)

Type to hold a second quantized Hamiltonian coefficients in memory

FermiCG.Molecule — Type

charge::Integer         #overall charge on molecule
multiplicity::Integer   #2S+1
atoms::Vector{Atom}     #Vector of `Atoms`
basis::String           #Basis set

Molecule essentially as a Vector of atoms, number of electrons and basis set

FermiCG.compute_energy — Method

compute_energy(h0, h1, h2, rdm1, rdm2)

Given an energy shift h0, 1e integrals h1, and 2e ints h2 along with a 1rdm and 2rdm on the same space, return the energy

FermiCG.compute_energy — Method

compute_energy(ints::InCoreInts, rdm1, rdm2)

Return energy defined by rdm1 and rdm2

FermiCG.orbital_rotation! — Method

orbital_rotation!(ints::InCoreInts, U)

Transform electronic integrals, by U i.e.,

\[h_{pq} = U_{rp}h_{rs}U_{sq}\]

\[(pq|rs) = (tu|vw)U_{tp}U_{uq}U_{vr}U_{ws}\]

FermiCG.orbital_rotation — Method

orbital_rotation(ints::InCoreInts, U)

Transform electronic integrals, by U i.e.,

\[h_{pq} = U_{rp}h_{rs}U_{sq}\]

\[(pq|rs) = (tu|vw)U_{tp}U_{uq}U_{vr}U_{ws}\]

FermiCG.subset — Function

subset(ints::InCoreInts, list; rmd1a=nothing, rdm1b=nothing)

Extract a subset of integrals acting on orbitals in list, returned as InCoreInts type

ints::InCoreInts: Integrals for full system

list: list of orbital indices in subset

rdm1a: 1RDM for embedding α density to make CASCI hamiltonian

rdm1b: 1RDM for embedding β density to make CASCI hamiltonian