FermiCG.get_ovlp

FermiCG.localize

FermiCG.make_pyscf_mole

FermiCG.pyscf_build_1e

FermiCG.pyscf_build_eri

FermiCG.pyscf_build_ints

FermiCG.pyscf_do_scf

FermiCG.pyscf_fci

FermiCG.pyscf_get_jk

FermiCG.pyscf_write_molden

FermiCG.pyscf_write_molden

PySCF Functions

FermiCG.get_ovlp — Method

get_ovlp(mf)

Get overlap matrix from pyscf using mean-field object

FermiCG.localize — Method

localize(C::Array{Float64,2},method::String, mf)

Localize the orbitals using method = method

FermiCG.make_pyscf_mole — Method

make_pyscf_mole(molecule::Molecule)

Create a pyscf.gto.Mole() object

FermiCG.pyscf_build_1e — Method

pyscf_build_1e(mol::Molecule)

build 1 electron integrals in AO basis

Arguments

mol::Molecule

returns a 2D matrix

FermiCG.pyscf_build_eri — Method

pyscf_build_eri(mol::Molecule, c1::Matrix, c2::Matrix, c3::Matrix, c4::Matrix)

build 2 electron integrals between different orbital spaces, (c1c2|c3c4)

Arguments

mol::Molecule

c1: active space orbital MO coeffs for index 1

c2: active space orbital MO coeffs for index 2

c3: active space orbital MO coeffs for index 3

c4: active space orbital MO coeffs for index 4

returns a 4D tensor

FermiCG.pyscf_build_ints — Method

pyscf_build_ints(mol, c_act, d1_embed)

build 1 and 2 electron integrals using a pyscf SCF object

Arguments

mol: PySCF Molecule object

c_act: active space orbital MO coeffs

d1_embed: 1rdm density matrix for the frozen part in the AO basis (e.g, doccs or frozen clusters)

returns an InCoreInts type

FermiCG.pyscf_do_scf — Method

pyscf_do_scf(molecule::Molecule, conv_tol=1e-10)

Use PySCF to compute Hartree-Fock for a given molecule and basis set and return a PYSCF mean field object

FermiCG.pyscf_fci — Method

pyscf_fci(ham, na, nb; max_cycle=20, conv_tol=1e-8, nroots=1, verbose=1)

Use PySCF to compute Full CI

FermiCG.pyscf_get_jk — Method

pyscf_get_jk(mol, density)

Build exchange matrix in AO basis

Arguments

mol: PySCF Molecule object

density: 1rdm density matrix for

FermiCG.pyscf_write_molden — Method

pyscf_write_molden(mf; filename="orbitals.molden")

Arguments

mf: PySCF mean field object

filename: Filename to write to

Write MO coeffs C to a molden file for visualizing

FermiCG.pyscf_write_molden — Method

pyscf_write_molden(molecule::Molecule, C; filename="orbitals.molden")

Arguments

molecule::Molecule: Molecule object

C: MO Coefficients

filename: Filename to write to

Write MO coeffs C to a molden file for visualizing