Wavefunction types

Each of the methods implemented in this package use a unique data representation for the wavefunction. These all rely on some kind of hash table lookup, and so the following indexing schemes are defined to make this easier and faster (i.e., implemented using only stack allocated data).

List of Indexing types for representing states

ClusterConfig

TuckerConfig

FockConfig

Each of the different wavefunction types map different Index types to data (wavefunction coefficients)

TPSCI: ClusteredState: Maps FockConfig → ClusterConfig → Vector of coefficients (one for each state)

BST: CompressedTuckerState: Maps FockConfig → TuckerConfig → Tucker instance representing compressed set of coefficients for that subspace

Index

Base.:+

Base.:-

Base.:-

Types

Methods

Base.:+ — Method

Base.:+(a::FockConfig, b::TransferConfig)

Add a FockConfig to a TransferConfig to get a new FockConfig

Base.:- — Method

Base.:-(a::FockConfig, b::TransferConfig)

Base.:- — Method

Base.:-(a::FockConfig, b::FockConfig)

Subtract two FockConfig's, returning a TransferConfig