FermiCG.StringCI

Base.copy!

FermiCG.StringCI.apply_annihilation!

FermiCG.StringCI.apply_creation!

FermiCG.StringCI.build_H_matrix

FermiCG.StringCI.build_S2_matrix

FermiCG.StringCI.calc_linear_index

FermiCG.StringCI.calc_linear_index!

FermiCG.StringCI.calc_linear_index!

FermiCG.StringCI.calc_nchk

FermiCG.StringCI.compute_AA

FermiCG.StringCI.compute_AAa

FermiCG.StringCI.compute_AB

FermiCG.StringCI.compute_ABa

FermiCG.StringCI.compute_ABb

FermiCG.StringCI.compute_Aa

FermiCG.StringCI.compute_Ab

FermiCG.StringCI.compute_ab_terms

FermiCG.StringCI.compute_ab_terms

FermiCG.StringCI.compute_ab_terms2

FermiCG.StringCI.compute_ab_terms_full

FermiCG.StringCI.compute_ab_terms_full!

FermiCG.StringCI.compute_annihilation

FermiCG.StringCI.compute_fock_diagonal

FermiCG.StringCI.compute_ss_terms2

FermiCG.StringCI.fill_ca_lookup

FermiCG.StringCI.fill_ca_lookup2

FermiCG.StringCI.fill_ca_lookup3

FermiCG.StringCI.get_map

FermiCG.StringCI.get_map

FermiCG.StringCI.get_nchk

FermiCG.StringCI.get_unoccupied

FermiCG.StringCI.get_unoccupied!

FermiCG.StringCI.precompute_spin_diag_terms

FermiCG.StringCI.reset!

FermiCG.StringCI.run_fci

FermiCG.StringCI.svd_state

StringCI

FermiCG.StringCI — Module

Simple string-based CI code

Base.copy! — Method

Base.copy!(c1::DeterminantString,c2::DeterminantString)

copy c2 into c1

FermiCG.StringCI.apply_annihilation! — Method

apply_annihilation!(c::DeterminantString, orb_index::Int)

Apply an annihilation operator to c corresponding to orbital orb_index

FermiCG.StringCI.apply_creation! — Method

apply_creation!(c::DeterminantString, orb_index::Int)

Apply a creation operator to c corresponding to orbital orb_index

FermiCG.StringCI.calc_linear_index! — Method

calc_linear_index!(c::DeterminantString, binomcoeff::Array{Int,2})

Calculate the linear index, passing in binomial coefficient matrix makes it much faster

FermiCG.StringCI.calc_linear_index! — Method

calc_linear_index!(c::DeterminantString)

Calculate the linear index

FermiCG.StringCI.calc_linear_index — Method

calc_linear_index(c::DeterminantString)

Return linear index for lexically ordered __config DeterminantString

FermiCG.StringCI.fill_ca_lookup — Method

fill_ca_lookup(c::DeterminantString)

Create an index table relating each DeterminantString with all ia substitutions i.e., calookup[Ka][c(p) + a(p)*np] = La

FermiCG.StringCI.fill_ca_lookup2 — Method

fill_ca_lookup2(c::DeterminantString)

Create an index table relating each DeterminantString with all ia substitutions i.e., ca_lookup[Ka,p,q] = sign*La

<L|p'q|K> = sign

FermiCG.StringCI.fill_ca_lookup3 — Method

fill_ca_lookup3(c::DeterminantString)

Create an index table relating each DeterminantString with all ia substitutions i.e., ca_lookup[Ka,p,q] = (sign,La)

<L|p'q|K> = sign

FermiCG.StringCI.get_unoccupied! — Method

get_unoccupied(c::DeterminantString)

get list of orbitals that are unoccupied

FermiCG.StringCI.get_unoccupied — Method

get_unoccupied(c::DeterminantString)

get list of orbitals that are unoccupied

FermiCG.StringCI.reset! — Method

reset!(c::DeterminantString)

Reset the DeterminantString to the first config

FermiCG.StringCI.build_H_matrix — Method

build_H_matrix(ints, P::FCIProblem)

Build the Hamiltonian defined by ints in the Slater Determinant Basis in the sector of Fock space specified by P

FermiCG.StringCI.build_S2_matrix — Method

build_s2(prb::FCIProblem)

prb: FCIProblem just defines the current CI problem (i.e., fock sector)

FermiCG.StringCI.compute_ab_terms — Method

compute_ab_terms(v, H, P::FCIProblem, ket_a_lookup, ket_b_lookup)

FermiCG.StringCI.compute_ab_terms — Method

compute_ab_terms(v, H, P::FCIProblem)

FermiCG.StringCI.compute_ab_terms2 — Method

compute_ab_terms2(v, H, P::FCIProblem, 
                      ket_a_lookup, ket_b_lookup)

FermiCG.StringCI.compute_ab_terms_full! — Method

compute_ab_terms_full!(H, P::FCIProblem, Hmat)

FermiCG.StringCI.compute_ab_terms_full — Method

compute_ab_terms_full(H, P::FCIProblem)

FermiCG.StringCI.compute_fock_diagonal — Method

compute_fock_diagonal!(H, P::FCIProblem, e_mf::Float64)

FermiCG.StringCI.compute_ss_terms2 — Method

compute_aa_terms2(v, H, P::FCIProblem, 
                      ket_a_lookup)

FermiCG.StringCI.get_map — Method

get_map(ham, prb::FCIProblem, HdiagA, HdiagB)

Assumes you've already computed the spin diagonal components

FermiCG.StringCI.get_map — Method

get_map(ham, prb::FCIProblem)

Get LinearMap with takes a vector and returns action of H on that vector

FermiCG.StringCI.precompute_spin_diag_terms — Method

precompute_spin_diag_terms(H, P::FCIProblem, e)

FermiCG.StringCI.run_fci — Method

run_fci(ints, prb::FCIProblem)

input: ints is a struct containing 0, 2, and 4 dimensional tensors

h0: energy shift

h1: 1 electron integrals

h2: 2 electron integrals (chemists notation)

prb: FCIProblem just defines the current CI problem (i.e., fock sector)

ints is simply an InCoreInts object from FermiCG

FermiCG.StringCI.svd_state — Method

svd_state(prb::FCIProblem)

Do an SVD of the FCI vector partitioned into clusters with (norbs1 | norbs2) where the orbitals are assumed to be ordered for cluster 1| cluster 2 haveing norbs1 and norbs2, respectively.

prb: FCIProblem just defines the current CI problem (i.e., fock sector)

norbs1:number of orbitals in left cluster

norbs2:number of orbitals in right cluster

svd_thresh: the threshold below which the states will be discarded

FermiCG.StringCI.calc_nchk — Method

calc_nchk(n::Integer,k::Integer)

Calculates binomial coefficient: n choose k

FermiCG.StringCI.get_nchk — Method

get_nchk(n::Integer,k::Integer)

Looks up binomial coefficient from a precomputed table: n choose k

FermiCG.StringCI.compute_AA — Method

compute_AA(no::Integer, bra_na, bra_nb, ket_na, ket_nb, bra_v::Matrix, ket_v::Matrix, spin_case)

Compute represntation of 2creation operators between states bra_v and ket_v

Arguments

no: number of orbitals

bra_na: number of alpha electrons in bra

bra_nb: number of beta electrons in bra

ket_na: number of alpha electrons in ket

ket_nb: number of beta electrons in ket

bra_v: basis vectors in bra

ket_v: basis vectors ket

spin_case: either (alpha|beta)

FermiCG.StringCI.compute_AAa — Method

compute_AAa(no::Integer, bra_na, bra_nb, ket_na, ket_nb, bra_v::Matrix, ket_v::Matrix, spin_case)

Compute represntation of 2creation operators between states bra_v and ket_v

Arguments

no: number of orbitals

bra_na: number of alpha electrons in bra

bra_nb: number of beta electrons in bra

ket_na: number of alpha electrons in ket

ket_nb: number of beta electrons in ket

bra_v: basis vectors in bra

ket_v: basis vectors ket

spin_case: either (alpha|beta)

FermiCG.StringCI.compute_AB — Method

compute_AB(no::Integer, bra_na, bra_nb, ket_na, ket_nb, bra_v::Matrix, ket_v::Matrix)

Compute representation of p'q' operators between states bra_v and ket_v, where p is alpha and q is beta.

Arguments

no: number of orbitals

bra_na: number of alpha electrons in bra

bra_nb: number of beta electrons in bra

ket_na: number of alpha electrons in ket

ket_nb: number of beta electrons in ket

bra_v: basis vectors in bra

ket_v: basis vectors ket

G(pqst) = v(IJs) <IJ|p'q'|KL> v(KLt) 
        = v(IJs) <J|<I|p'q'|K>|L> v(KLt)
        = v(IJs) <J|<I|p'|K>q'|L> v(KLt) (-1)^N(K)
        = v(IJs) <I|p'|K> <J|q'|L> v(KLt) (-1)^N(K)
    
        this can be vectorized if needed. for now:
        loop over p,q
            loop over K,L
                sign = -1^N(K)
                find indices of I,J
                G(pq,st) += v(IJ,s) v(KL,t) * sign 
            end
        end

FermiCG.StringCI.compute_ABa — Method

compute_ABa(no::Integer, bra_na, bra_nb, ket_na, ket_nb, bra_v::Matrix, ket_v::Matrix, spin_case)

Compute represntation of 2creation operators between states bra_v and ket_v

Arguments

no: number of orbitals

bra_na: number of alpha electrons in bra

bra_nb: number of beta electrons in bra

ket_na: number of alpha electrons in ket

ket_nb: number of beta electrons in ket

bra_v: basis vectors in bra

ket_v: basis vectors ket

spin_case: either (alpha|beta) giving ABa or BAb

FermiCG.StringCI.compute_ABb — Method

compute_ABb(no::Integer, bra_na, bra_nb, ket_na, ket_nb, bra_v::Matrix, ket_v::Matrix, spin_case)

Compute represntation of 2creation operators between states bra_v and ket_v

Arguments

no: number of orbitals

bra_na: number of alpha electrons in bra

bra_nb: number of beta electrons in bra

ket_na: number of alpha electrons in ket

ket_nb: number of beta electrons in ket

bra_v: basis vectors in bra

ket_v: basis vectors ket

spin_case: either (alpha|beta) giving ABa or BAb

FermiCG.StringCI.compute_Aa — Method

compute_Aa(no::Integer, bra_na, bra_nb, ket_na, ket_nb, bra_v::Matrix, ket_v::Matrix, spin_case)

Compute representation of a'a operators between states bra_v and ket_v

Arguments

no: number of orbitals

bra_na: number of alpha electrons in bra

bra_nb: number of beta electrons in bra

ket_na: number of alpha electrons in ket

ket_nb: number of beta electrons in ket

bra_v: basis vectors in bra

ket_v: basis vectors ket

spin_case: either (alpha|beta)

FermiCG.StringCI.compute_Ab — Method

compute_Ab(no::Integer, bra_na, bra_nb, ket_na, ket_nb, bra_v::Matrix, ket_v::Matrix)

Compute representation of p'q operators between states bra_v and ket_v, where p is alpha and q is beta.

Arguments

no: number of orbitals

bra_na: number of alpha electrons in bra

bra_nb: number of beta electrons in bra

ket_na: number of alpha electrons in ket

ket_nb: number of beta electrons in ket

bra_v: basis vectors in bra

ket_v: basis vectors ket

G(pqst) = v(IJs) <IJ|p'q|KL> v(KLt) 
        = v(IJs) <J|<I|p'q|K>|L> v(KLt)
        = v(IJs) <J|<I|p'|K>q|L> v(KLt) (-1)^N(K)
        = v(IJs) <I|p'|K> <J|q|L> v(KLt) (-1)^N(K)
    
        this can be vectorized if needed. for now:
        loop over p,q
            loop over K,L
                sign = -1^N(K)
                find indices of I,J
                G(pq,st) += v(IJ,s) v(KL,t) * sign 
            end
        end

FermiCG.StringCI.compute_annihilation — Method

compute_annihilation(no::Integer, bra_na, bra_nb, ket_na, ket_nb, bra_v::Matrix, ket_v::Matrix, spin_case)

Compute an creation operator between states bra_v and ket_v

Arguments

no: number of orbitals

bra_na: number of alpha electrons in bra

bra_nb: number of beta electrons in bra

ket_na: number of alpha electrons in ket

ket_nb: number of beta electrons in ket

bra_v: basis vectors in bra

ket_v: basis vectors ket

spin_case: either (alpha|beta)