FermiCG.build_compressed_1st_order_state

FermiCG.build_dense_H_term

FermiCG.build_sigma!

FermiCG.build_sigma_serial!

FermiCG.cache_hamiltonian

FermiCG.contract_dense_H_with_state

FermiCG.define_foi_space

FermiCG.get_map

FermiCG.hylleraas_compressed_mp2

FermiCG.project_out!

FermiCG.solve_for_compressed_space

FermiCG.tucker_cepa_solve

FermiCG.tucker_cepa_solve2

FermiCG.tucker_ci_solve

Tucker

FermiCG.build_dense_H_term — Method

Contract integrals and ClusterOps to form dense 4-body Hamiltonian matrix (tensor) in Tucker basis

FermiCG.build_sigma! — Method

build_sigma_parallel!(sigma_vector::CompressedTuckerState, ci_vector::CompressedTuckerState, cluster_ops, clustered_ham)

FermiCG.build_sigma_serial! — Method

build_sigma!(sigma_vector::CompressedTuckerState, ci_vector::CompressedTuckerState, cluster_ops, clustered_ham)

FermiCG.cache_hamiltonian — Method

build_sigma!(sigma_vector::CompressedTuckerState, ci_vector::CompressedTuckerState, cluster_ops, clustered_ham)

FermiCG.contract_dense_H_with_state — Method

Contract  Hamiltonian matrix (tensor) in Tucker basis with trial vector (Tucker)

FermiCG.build_compressed_1st_order_state — Method

build_compressed_1st_order_state(cts::CompressedTuckerState{T,N}, cluster_ops, clustered_ham; thresh=1e-7, max_number=nothing, nbody=2) where {T,N}

Apply the Hamiltonian to v expanding into the uncompressed space. This is done only partially, where each term is recompressed after being computed. Lots of overhead probably from compression, but never completely uncompresses.

#Arguments

cts::CompressedTuckerState: input state

cluster_ops:

clustered_ham: Hamiltonian

thresh: Threshold for Tucker decomosition

max_number: max number of tucker factors kept

nbody: allows one to limit (max 4body) terms in the Hamiltonian considered

#Returns

v1::CompressedTuckerState

FermiCG.define_foi_space — Method

define_foi_space(v::CompressedTuckerState, clustered_ham; nbody=2)

Compute the first-order interacting space as defined by clustered_ham

#Arguments

v::CompressedTuckerState: input state

clustered_ham: Hamiltonian

nbody: allows one to limit (max 4body) terms in the Hamiltonian considered

#Returns

foi::OrderedDict{FockConfig,Vector{TuckerConfig}}

FermiCG.get_map — Method

get_map(ci_vector::CompressedTuckerState, cluster_ops, clustered_ham)

Get LinearMap with takes a vector and returns action of H on that vector

FermiCG.hylleraas_compressed_mp2 — Method

hylleraas_compressed_mp2(sig_in::CompressedTuckerState, ref::CompressedTuckerState,
        cluster_ops, clustered_ham;
        H0 = "Hcmf", tol=1e-6, nbody=4, max_iter=40, verbose=1, do_pt = true, thresh=1e-8)

H0: ["H", "Hcmf"]

FermiCG.project_out! — Method

project_out!(v::CompressedTuckerState, w::CompressedTuckerState; thresh=1e-16)

Project w out of v |v'> = |v> - |w><w|v>

FermiCG.solve_for_compressed_space — Method

solve_for_compressed_space(ref_vector::CompressedTuckerState, cluster_ops, clustered_ham;
    max_iter    = 20,
    H0          = "cmf",
    nbody       = 4,
    thresh_foi  = 1e-6,
    thresh_var  = 1e-4,
    tol_ci      = 1e-5,
    tol_tucker  = 1e-6)

Arguments

ref_vector: initial state

nbody: max number of nbody terms in the Hamiltonian used for creating FOIS

thresh_foi: Compression threshold for the FOIS, or first order wavefunction

thresh_var: Compression threshold for the variational solution

tol_ci: Convergence threshold for the CI (norm of residual)

tol_tucker: Convergence threshold for Tucker iterations (energy change)

H0: ["Hcmf", "H"]

FermiCG.tucker_cepa_solve — Method

tucker_cepa_solve!(ref_vector::CompressedTuckerState, cepa_vector::CompressedTuckerState, cluster_ops, clustered_ham; tol=1e-5, cache=true)

Arguments

ref_vector: Input reference state.

cepa_vector: CompressedTuckerState which defines the configurational space defining {X}. This

should be the first-order interacting space (or some compressed version of it).

cluster_ops

clustered_ham

tol: haven't yet set this up (NYI)

cache: Should we cache the compressed H operators? Speeds up drastically, but uses lots of memory

Compute compressed CEPA. Since there can be non-zero overlap with a multireference state, we need to generalize.

HC = SCe

|Haa + Hax| |1 | = |I + Sax| |1 | E |Hxa + Hxx| |Cx| |Sxa + I | |Cx|

Haa + HaxCx = (1 + SaxCx)E Hxa + HxxCx = SxaE + CxE

The idea for CEPA is to approximate E in the amplitude equation. CEPA(0): E = Eref

(Hxx-Eref)Cx = SxaEref - Hxa

Ax=b

After solving, the Energy can be obtained as: E = (Eref + HaxCx) / (1 + SaxCx)

FermiCG.tucker_cepa_solve2 — Method

tucker_cepa_solve!(ref_vector::CompressedTuckerState, cepa_vector::CompressedTuckerState, cluster_ops, clustered_ham; tol=1e-5, cache=true)

Arguments

ref_vector: Input reference state.

cepa_vector: CompressedTuckerState which defines the configurational space defining {X}. This

should be the first-order interacting space (or some compressed version of it).

cluster_ops

clustered_ham

tol: haven't yet set this up (NYI)

cache: Should we cache the compressed H operators? Speeds up drastically, but uses lots of memory

Compute compressed CEPA. Since there can be non-zero overlap with a multireference state, we need to generalize.

HC = SCe

|Haa + Hax| |1 | = |I + Sax| |1 | E |Hxa + Hxx| |Cx| |Sxa + I | |Cx|

Haa + HaxCx = (1 + SaxCx)E Hxa + HxxCx = SxaE + CxE

The idea for CEPA is to approximate E in the amplitude equation. CEPA(0): E = Eref

(Hxx-Eref)Cx = SxaEref - Hxa

Ax=b

After solving, the Energy can be obtained as: E = (Eref + HaxCx) / (1 + SaxCx)

FermiCG.tucker_ci_solve — Method

tucker_ci_solve(ci_vector::CompressedTuckerState, cluster_ops, clustered_ham; tol=1e-5)

Solve for ground state in the space spanned by ci_vector's compression vectors