BST

The Block-Sparse Tucker (BST) method, approximates FCI as a linear combination of individually compressed (via HOSVD) blocks of the Hilbert space.

Background

Similar to TPSCI, the approach here starts with a CMF wavefunction, and systematically reintroduces the discarded tensor product states to variationally approach FCI. Unlike with TPSCI, however, we don't assume that the final wavefunction is written as a purely sparse form (where only a few TPS's are needed), but rather we assume that collections of TPS's where certain numbers of clusters are "excited" can be efficiently compressed via HOSVD (although the basic idea would extend to other tensor decompositions, like CP or MPS).

Performance considerations

Index

FermiCG.Tucker

FermiCG.add

FermiCG.block_sparse_tucker

FermiCG.build_compressed_1st_order_state

FermiCG.build_dense_H_term

FermiCG.build_sigma!

FermiCG.build_sigma_serial!

FermiCG.cache_hamiltonian_old

FermiCG.compress

FermiCG.contract_dense_H_with_state

FermiCG.define_foi_space

FermiCG.do_fois_pt2

FermiCG.dot

FermiCG.get_map

FermiCG.hylleraas_compressed_mp2

FermiCG.project_out!

FermiCG.transform_basis

FermiCG.transform_basis

FermiCG.tucker_cepa_solve

FermiCG.tucker_cepa_solve2

FermiCG.tucker_ci_solve

FermiCG.tucker_decompose

FermiCG.tucker_decompose_tot

FermiCG.tucker_recompose

Documentation

FermiCG.build_dense_H_term — Method

Contract integrals and ClusterOps to form dense 4-body Hamiltonian matrix (tensor) in Tucker basis

FermiCG.build_sigma! — Method

build_sigma_parallel!(sigma_vector::CompressedTuckerState, ci_vector::CompressedTuckerState, cluster_ops, clustered_ham)

FermiCG.build_sigma_serial! — Method

build_sigma!(sigma_vector::CompressedTuckerState, ci_vector::CompressedTuckerState, cluster_ops, clustered_ham)

FermiCG.cache_hamiltonian_old — Method

build_sigma!(sigma_vector::CompressedTuckerState, ci_vector::CompressedTuckerState, cluster_ops, clustered_ham)

FermiCG.contract_dense_H_with_state — Method

Contract  Hamiltonian matrix (tensor) in Tucker basis with trial vector (Tucker)

FermiCG.build_compressed_1st_order_state — Method

build_compressed_1st_order_state(ket_cts::CompressedTuckerState{T,N}, cluster_ops, clustered_ham; 
    thresh=1e-7, 
    max_number=nothing, 
    nbody=4) where {T,N}

Apply the Hamiltonian to v expanding into the uncompressed space. This is done only partially, where each term is recompressed after being computed. Lots of overhead probably from compression, but never completely uncompresses.

#Arguments

cts::CompressedTuckerState: input state

cluster_ops:

clustered_ham: Hamiltonian

thresh: Threshold for each HOSVD

max_number: max number of tucker factors kept in each HOSVD

nbody: allows one to limit (max 4body) terms in the Hamiltonian considered

#Returns

v1::CompressedTuckerState

FermiCG.define_foi_space — Method

define_foi_space(v::CompressedTuckerState, clustered_ham; nbody=2)

Compute the first-order interacting space as defined by clustered_ham

#Arguments

v::CompressedTuckerState: input state

clustered_ham: Hamiltonian

nbody: allows one to limit (max 4body) terms in the Hamiltonian considered

#Returns

foi::OrderedDict{FockConfig,Vector{TuckerConfig}}

FermiCG.do_fois_pt2 — Method

function do_fois_pt2(ref::CompressedTuckerState, cluster_ops, clustered_ham;
        H0          = "Hcmf",
        max_iter    = 50,
        nbody       = 4,
        thresh_foi  = 1e-6,
        tol         = 1e-5,
        opt_ref     = true,
        verbose     = true)

Do PT2

FermiCG.get_map — Method

get_map(ci_vector::CompressedTuckerState, cluster_ops, clustered_ham)

Get LinearMap with takes a vector and returns action of H on that vector

FermiCG.hylleraas_compressed_mp2 — Method

hylleraas_compressed_mp2(sig_in::CompressedTuckerState, ref::CompressedTuckerState,
        cluster_ops, clustered_ham;
        H0 = "Hcmf", tol=1e-6, nbody=4, max_iter=40, verbose=1, do_pt = true, thresh=1e-8)

H0: ["H", "Hcmf"]

FermiCG.project_out! — Method

project_out!(v::CompressedTuckerState, w::CompressedTuckerState; thresh=1e-16)

Project w out of v |v'> = |v> - |w><w|v>

FermiCG.tucker_cepa_solve — Function

tucker_cepa_solve!(ref_vector::CompressedTuckerState, cepa_vector::CompressedTuckerState, cluster_ops, clustered_ham; tol=1e-5, cache=true)

Arguments

ref_vector: Input reference state.

cepa_vector: CompressedTuckerState which defines the configurational space defining {X}. This

should be the first-order interacting space (or some compressed version of it).

cluster_ops

clustered_ham

tol: haven't yet set this up (NYI)

cache: Should we cache the compressed H operators? Speeds up drastically, but uses lots of memory

Compute compressed CEPA. Since there can be non-zero overlap with a multireference state, we need to generalize.

HC = SCe

|Haa + Hax| |1 | = |I   + Sax| |1 | E
|Hxa + Hxx| |Cx|   |Sxa + I  | |Cx|

Haa + Hax*Cx = (1 + Sax*Cx)E
Hxa + HxxCx = SxaE + CxE

The idea for CEPA is to approximate E in the amplitude equation. CEPA(0): E = Eref

(Hxx-Eref)*Cx = Sxa*Eref - Hxa

Ax=b

After solving, the Energy can be obtained as:

E = (Eref + Hax*Cx) / (1 + Sax*Cx)

FermiCG.tucker_cepa_solve2 — Method

tucker_cepa_solve!(ref_vector::CompressedTuckerState, cepa_vector::CompressedTuckerState, cluster_ops, clustered_ham; tol=1e-5, cache=true)

Arguments

ref_vector: Input reference state.

cepa_vector: CompressedTuckerState which defines the configurational space defining {X}. This

should be the first-order interacting space (or some compressed version of it).

cluster_ops

clustered_ham

tol: haven't yet set this up (NYI)

cache: Should we cache the compressed H operators? Speeds up drastically, but uses lots of memory

Compute compressed CEPA. Since there can be non-zero overlap with a multireference state, we need to generalize.

HC = SCe

|Haa + Hax| |1 | = |I + Sax| |1 | E |Hxa + Hxx| |Cx| |Sxa + I | |Cx|

Haa + HaxCx = (1 + SaxCx)E Hxa + HxxCx = SxaE + CxE

The idea for CEPA is to approximate E in the amplitude equation. CEPA(0): E = Eref

(Hxx-Eref)Cx = SxaEref - Hxa

Ax=b

After solving, the Energy can be obtained as: E = (Eref + HaxCx) / (1 + SaxCx)

FermiCG.tucker_ci_solve — Method

tucker_ci_solve(ci_vector::CompressedTuckerState, cluster_ops, clustered_ham; tol=1e-5)

Solve for ground state in the space spanned by ci_vector's compression vectors

FermiCG.block_sparse_tucker — Method

block_sparse_tucker(input_vec::CompressedTuckerState, cluster_ops, clustered_ham;
    max_iter    = 20,
    max_iter_pt = 200, 
    nbody       = 4,
    H0          = "Hcmf",
    thresh_var  = 1e-4,
    thresh_foi  = 1e-6,
    thresh_pt   = 1e-5,
    tol_ci      = 1e-5,
resolve_ss  = true,
    do_pt       = true,
    tol_tucker  = 1e-6)

Arguments

input_vec::CompressedTuckerState: initial state

cluster_ops: local cluster operators

clustered_ham::ClusteredOperator: hamiltonian

max_iter = 20: max number of iterations

max_iter_pt = 200: max number of iterations to solve for the 1st order wavefunction

nbody = 4: include up to nbody terms when forming the FOIS to search

H0 = "Hcmf": zeroth-order hamiltonian for computing 1st order wavefunction ["Hcmf", "H"]

nbody: max number of nbody terms in the Hamiltonian used for creating FOIS

thresh_var: Compression threshold for the variational solution

thresh_foi: Compression threshold for the FOIS

thresh_pt: Compression threshold for the first-order wavefunction (if used)

tol_ci: Convergence threshold for the CI (norm of residual)

resolve_ss: After compressing previous variational state, should we resolve in new subspace?

do_pt = true: Compute pt1 wavefunction for finding updated compression basis?

tol_tucker: Convergence threshold for Tucker iterations (energy change)

Returns

e_var::Float64: the final variational energy

v_var::CompressedTuckerState: the final variational state

See also: CompressedTuckerState, Tucker

HOSVD

FermiCG.Tucker — Type

Simple Tucker (HOSVD) type

Data

core::Array{T, N}

factors::NTuple{N, Matrix{T}}

Tucker factors are stored as tall matrices

FermiCG.add — Method

add(tucks::Vector{Tucker{T,N}}; thresh=1e-10, max_number=nothing) where {T,N}

Add together multiple Tucker instances. Assumed non-orthogonal.

Arguments

tucks::Vector{Tucker{T,N}}: Vector of Tucker objects

FermiCG.compress — Method

compress(t::Tucker{T,N}; thresh=1e-7, max_number=nothing) where {T,N}

Try to compress further

FermiCG.dot — Method

dot(t1::Tucker{T,N}, t2::Tucker{T,N}) where {T,N}

Note: This doesn't assume t1 and t2 have the same compression vectors

FermiCG.transform_basis — Method

FermiCG.transform_basis — Method

FermiCG.tucker_decompose — Method

function tucker_decompose(A::Array{T,N}; thresh=1e-7, max_number=nothing, verbose=1, type="magnitude") where {T,N}

Tucker Decomposition of dense tensor: A ~ X *(1) U1 *(2) U2 .... where cluster states are discarded based on the corresponding SVD #Arguments

A: matrix to decompose

thresh: threshold for discarding tucker factors

max_number: limit number of tucker factors to this value

type: type of trunctation. "magnitude" discards values smaller than this number. "sum" discards values such that the sum of discarded values is smaller than thresh.

FermiCG.tucker_decompose_tot — Method

function tucker_decompose_tot(A::Array{T,N}; thresh=1e-7, verbose=1) where {T,N}

Tucker Decomposition of dense tensor: A ~ X *(1) U1 *(2) U2 .... where the cluster states are discarded to ensure ||V-v||_F < thresh, with V and v being the full and approximated tensors. This isn't quite ready for use, as it doens't fall back to the optimal case for SVD of a matrix as it doesn't consider the possibility of being diagonal in svd basis.

FermiCG.tucker_recompose — Method

tucker_recompose(core, factors)

Recompose Tucker Decomposition