Wavefunction types

Each of the methods implemented in this package use a unique data representation for the wavefunction. These all rely on some kind of hash table lookup, and so the following indexing schemes are defined to make this easier and faster (i.e., implemented using only stack allocated data).

List of Indexing types for representing states

ClusterConfig

TuckerConfig

FockConfig

Each of the different wavefunction types map different Index types to data (wavefunction coefficients)

TPSCI: ClusteredState: Maps FockConfig → ClusterConfig → Vector of coefficients (one for each state)

BST: CompressedTuckerState: Maps FockConfig → TuckerConfig → Tucker instance representing compressed set of coefficients for that subspace

Index

FermiCG.AbstractState

FermiCG.ClusterConfig

FermiCG.ClusteredState

FermiCG.ClusteredState

FermiCG.CompressedTuckerState

FermiCG.CompressedTuckerState

FermiCG.FockConfig

FermiCG.FockSparseState

FermiCG.OperatorConfig

FermiCG.SparseIndex

FermiCG.TransferConfig

FermiCG.TuckerConfig

FermiCG.TuckerState

FermiCG.TuckerState

FermiCG.TuckerState

Base.:+

Base.:+

Base.:-

Base.Multimedia.display

Base.Multimedia.display

Base.Multimedia.display

Base.convert

Base.convert

Base.convert

Base.convert

Base.length

Base.length

Base.setindex!

FermiCG.add!

FermiCG.add!

FermiCG.add_fockconfig!

FermiCG.add_fockconfig!

FermiCG.add_fockconfig!

FermiCG.add_fockconfig!

FermiCG.clip!

FermiCG.compress

FermiCG.dim

FermiCG.dim

FermiCG.dot

FermiCG.dot

FermiCG.dot

FermiCG.expand_each_fock_space!

FermiCG.eye!

FermiCG.fold!

FermiCG.get_vector

FermiCG.get_vector

FermiCG.get_vector

FermiCG.get_vectors

FermiCG.is_subset

FermiCG.mult!

FermiCG.nonorth_add!

FermiCG.nonorth_dot

FermiCG.normalize!

FermiCG.orth!

FermiCG.orth_add!

FermiCG.orth_dot

FermiCG.orthogonalize!

FermiCG.print_configs

FermiCG.print_fock_occupations

FermiCG.print_fock_occupations

FermiCG.prune_empty_TuckerConfigs!

FermiCG.prune_empty_fock_spaces!

FermiCG.prune_empty_fock_spaces!

FermiCG.prune_empty_fock_spaces!

FermiCG.randomize!

FermiCG.replace

FermiCG.replace

FermiCG.scale!

FermiCG.scale!

FermiCG.scale!

FermiCG.set_vector!

FermiCG.set_vector!

FermiCG.set_vector!

FermiCG.unfold!

FermiCG.zero!

FermiCG.zero!

FermiCG.zero!

LinearAlgebra.norm

LinearAlgebra.norm

Types

FermiCG.AbstractState — Type

Glue different Sparse Vector State types together

FermiCG.FockSparseState — Type

FermiCG.ClusteredState — Type

clusters::Vector{Cluster}
data::OrderedDict{FockConfig{N}, OrderedDict{ClusterConfig{N}, Vector{T}}}

This represents an arbitrarily sparse state. E.g., used in TPSCI

FermiCG.ClusteredState — Method

ClusteredState(clusters)

Constructor

clusters::Vector{Cluster}

FermiCG.TuckerState — Type

Represents a state in an set of abitrary (yet dense) subspaces of a set of FockConfigs.

v[FockConfig][TuckerConfig] => Dense Tensor E.g., used in n-body Tucker

clusters::Vector{Cluster}
data::OrderedDict{FockConfig,OrderedDict{TuckerConfig,Array}}
p_spaces::Vector{ClusterSubspace}
q_spaces::Vector{ClusterSubspace}

FermiCG.TuckerState — Method

TuckerState(clusters, p_spaces, q_spaces, foi; nroots=1)

Constructor

clusters::Vector{Cluster}

p_spaces::Vector{ClusterSubspace}

q_spaces::Vector{ClusterSubspace}

na::Int Number of alpha

nb::Int Number of beta

FermiCG.TuckerState — Method

TuckerState(clusters, p_spaces, q_spaces, foi; nroots=1)

Constructor to build state directly from a definition of a first order interacting space (or more generic even i suppose)

clusters::Vector{Cluster}

p_spaces::Vector{ClusterSubspace}

q_spaces::Vector{ClusterSubspace}

foi::OrderedDict{FockConfig,Vector{TuckerConfig}}

nroots

FermiCG.CompressedTuckerState — Type

Represents a state in an set of abitrary (yet low-rank) subspaces of a set of FockConfigs. e.g.

v[FockConfig][TuckerConfig] => Tucker Decomposed Tensor

Data

clusters::Vector{Cluster}

data::OrderedDict{FockConfig,OrderedDict{TuckerConfig,Tucker}}

p_spaces::Vector{ClusterSubspace}

q_spaces::Vector{ClusterSubspace}

FermiCG.CompressedTuckerState — Method

CompressedTuckerState(ts::TuckerState; thresh=-1, max_number=nothing, verbose=0)

Create a CompressedTuckerState from a TuckerState

Arguments

ts::TuckerState

thresh=-1: discard singular values smaller than thresh

max_number=nothing: if != nothing, only keep up to max_number singular vectors per SVD

verbose=0: print level

Returns

CompressedTuckerState

FermiCG.ClusterConfig — Type

config::NTuple{N,Int16}

Indexes an N dimensional space

FermiCG.FockConfig — Type

config::NTuple{N,Tuple{Int16,Int16}}

Indexes a 'Change in fock space'. For instance, if F1 and F2 are distince FockConfig instances, then TransferConfig is F1-F2.

e.g., config = ((2,3), (2,2), (3,2)) is a 3 cluster configuration with 2, 2, 3 α and 3, 2, 2 β electrons, respectively.

FermiCG.OperatorConfig — Type

config::Tuple{FockConfig{N}, FockConfig{N}, T, T}

FermiCG.SparseIndex — Type

Abstract index type

FermiCG.TransferConfig — Type

config::NTuple{N,Tuple{Int16,Int16}}

Indexes a 'Change in fock space'. For instance, if F1 and F2 are distince FockConfig instances, then TransferConfig is F1-F2.

This is mainly used to label individual parts of an operator (Hamiltonian) as they can only enact certain changes in the Fock space configurations.

FermiCG.TuckerConfig — Type

config::NTuple{N,UnitRange{Int}}

Indexes a particular subspace in the tensor product space. Each TuckerConfig instance specifies the entire subspace defined by a set of cluster states (specified by a range) on each cluster.

E.g.,

((1:1), (1:1), (1:1)) is simply the ground state CMF state for a 3 cluster systems.

((2:20), (1:1), (2:20)) are all states where clusters 1 and 2 are excited out of their repsective ground states.

Methods

Base.setindex! — Method

setindex!(s::AbstractState, a::OrderedDict, b)

FermiCG.add_fockconfig! — Method

add_fockconfig!(s::AbstractState, fock::Vector{Tuple{T,T}}) where T<:Integer

Base.Multimedia.display — Method

Base.display(s::ClusteredState; thresh=1e-3, root=1)

Pretty print

FermiCG.add! — Method

add!(s1::ClusteredState, s2::ClusteredState)

Add coeffs in s2 to s1

FermiCG.add_fockconfig! — Method

add_fockconfig!(s::ClusteredState, fock::FockConfig)

FermiCG.clip! — Method

clip!(s::ClusteredState; thresh=1e-5)

FermiCG.dot — Method

dot(v1::ClusteredState,v2::ClusteredState; r1=1, r2=1)

FermiCG.dot — Method

dot(v1::ClusteredState,v2::ClusteredState; r1=1, r2=1)

FermiCG.get_vector — Method

get_vector(s::ClusteredState; root=1)

FermiCG.get_vectors — Method

get_vectors(s::ClusteredState)

FermiCG.normalize! — Method

normalize!(s::AbstractState)

FermiCG.orth! — Method

dot(v1::ClusteredState,v2::ClusteredState; r1=1, r2=1)

FermiCG.print_configs — Method

print_configs(s::ClusterState; thresh=1e-3)

Pretty print

FermiCG.prune_empty_fock_spaces! — Method

prune_empty_fock_spaces!(s::ClusteredState)

remove fock_spaces that don't have any configurations

FermiCG.scale! — Method

scale!(s::ClusteredState,c)

FermiCG.set_vector! — Method

set_vector!(s::ClusteredState)

FermiCG.zero! — Method

zero!(s::ClusteredState)

set all elements to zero

LinearAlgebra.norm — Method

norm(s::ClusteredState, root)

LinearAlgebra.norm — Method

norm(s::ClusteredState{T,N,R}) where {T,N,R}

Base.:+ — Method

+(ts1::FermiCG.TuckerState, ts2::FermiCG.TuckerState)

Base.Multimedia.display — Method

Base.display(s::TuckerState; thresh=1e-3)

Pretty print

Base.length — Method

Base.length(s::TuckerState)

FermiCG.add! — Method

add!(ts1::FermiCG.TuckerState, ts2::FermiCG.TuckerState)

Add coeffs in ts2 to ts1

FermiCG.add_fockconfig! — Method

add_fockconfig!(s::ClusteredState, fock::FockConfig)

FermiCG.dot — Method

dot(ts1::FermiCG.TuckerState, ts2::FermiCG.TuckerState)

Dot product between in ts2 to ts1

FermiCG.expand_each_fock_space! — Method

expand_each_fock_space!(s::TuckerState, bases)

For each fock space sector defined, add all possible basis states

bases::Vector{ClusterBasis}

FermiCG.eye! — Method

eye!(s::TuckerState)

FermiCG.fold! — Method

fold!(ts::TuckerState)

FermiCG.get_vector — Method

get_vector(s::TuckerState)

FermiCG.mult! — Method

mult!(ts::TuckerState, A)

Multiple ts by a matrix A. This is a multiplication over global state index

FermiCG.orthogonalize! — Method

orthogonalize!(ts::TuckerState)

Symmetric Orthogonalization of vectors in ts

FermiCG.print_fock_occupations — Method

print_fock_occupations(s::TuckerState; root=1, thresh=1e-3)

Pretty print

FermiCG.prune_empty_fock_spaces! — Method

prune_empty_fock_spaces!(s::TuckerState)

remove fock_spaces that don't have any configurations

FermiCG.randomize! — Method

randomize!(ts::TuckerState; scale=1)

Add some random noise to the vector

FermiCG.scale! — Method

scale!(ts::FermiCG.TuckerState, a)

Scale ts by a constant a

FermiCG.set_vector! — Method

set_vector!(s::TuckerState)

FermiCG.unfold! — Method

unfold!(ts::TuckerState)

FermiCG.zero! — Method

zero!(s::TuckerState)

Base.Multimedia.display — Method

Base.display(s::CompressedTuckerState; thresh=1e-3)

Pretty print

Base.length — Method

Base.length(s::CompressedTuckerState)

FermiCG.add_fockconfig! — Method

add_fockconfig!(s::CompressedTuckerState, fock::FockConfig)

FermiCG.compress — Method

compress(ts::CompressedTuckerState{T,N}; thresh=-1, max_number=nothing, verbose=0) where {T,N}

Compress state via HOSVD

Arguments

ts::CompressedTuckerState

thresh = -1: threshold for compression

max_number: only keep certain number of vectors per TuckerConfig

verbose=0: print level

Returns

CompressedTuckerState

FermiCG.get_vector — Method

get_vector(s::CompressedTuckerState)

Return a vector of the variables. Note that this is the core tensors being returned

FermiCG.nonorth_add! — Method

nonorth_add!(ts1::CompressedTuckerState, ts2::CompressedTuckerState)

Add coeffs in ts2 to ts1

Note: this does not assume t1 and t2 have the same compression vectors

FermiCG.nonorth_dot — Method

nonorth_dot(ts1::FermiCG.CompressedTuckerState, ts2::FermiCG.CompressedTuckerState; verbose=0)

Dot product between 1ts2andts1` where each have their own Tucker factors

FermiCG.orth_add! — Method

orth_add!(ts1::CompressedTuckerState, ts2::CompressedTuckerState)

Add coeffs in ts2 to ts1

Note: this assumes t1 and t2 have the same compression vectors

FermiCG.orth_dot — Method

dot(ts1::FermiCG.CompressedTuckerState, ts2::FermiCG.CompressedTuckerState)

Dot product between ts2 and ts1

Warning: this assumes both ts1 and ts2 have the same tucker factors for each TuckerConfig

FermiCG.print_fock_occupations — Method

print_fock_occupations(s::CompressedTuckerState; thresh=1e-3)

Pretty print

FermiCG.prune_empty_TuckerConfigs! — Method

prune_empty_TuckerConfigs!(s::T) where T<:Union{TuckerState, CompressedTuckerState}

remove fock_spaces that don't have any configurations

FermiCG.prune_empty_fock_spaces! — Method

prune_empty_fock_spaces!(s::AbstractState)

remove fock_spaces that don't have any configurations

FermiCG.scale! — Method

scale!(ts::FermiCG.CompressedTuckerState, a::T<:Number)

Scale ts by a constant

FermiCG.set_vector! — Method

set_vector!(s::CompressedTuckerState)

FermiCG.zero! — Method

zero!(s::CompressedTuckerState)

Base.:+ — Method

Base.:+(a::FockConfig, b::TransferConfig)

Add a FockConfig to a TransferConfig to get a new FockConfig

Base.:- — Method

Base.:-(a::FockConfig, b::FockConfig)

Subtract two FockConfig's, returning a TransferConfig

Base.convert — Method

function Base.convert(::Type{TransferConfig{N}}, in::NTuple{M,Tuple{T,T}}) where {T,N,M}

Base.convert — Method

function Base.convert(::Type{FockConfig{N}}, in::Vector) where {N}

Base.convert — Method

function Base.convert(::Type{TransferConfig{N}}, in::Vector{Tuple{T,T}}) where {T,N}

Base.convert — Method

function Base.convert(::Type{TuckerConfig{N}}, in::Vector{UnitRange{T}}) where {T,N}

FermiCG.dim — Method

dim(tc::TuckerConfig)

Return total dimension of space indexed by tc

FermiCG.dim — Method

dim(fc::FockConfig, no)

Return total dimension of space indexed by fc on no orbitals

FermiCG.is_subset — Method

Check if tc1 is a subset of tc2

FermiCG.replace — Method

function replace(tc::TransferConfig, idx, fock)

FermiCG.replace — Method

function replace(cc::ClusterConfig{N}, idx, conf) where N