FermiCG.build_brdm

FermiCG.build_full_H

FermiCG.build_full_H_parallel

FermiCG.compute_diagonal

FermiCG.compute_expectation_value

FermiCG.compute_pt2

FermiCG.contract_matrix_element

FermiCG.contract_matrix_element

FermiCG.contract_matrix_element

FermiCG.contract_matrix_element

FermiCG.contract_matvec

FermiCG.contract_matvec

FermiCG.contract_matvec

FermiCG.contract_matvec

FermiCG.contract_matvec_M3

FermiCG.contract_matvec_M4

FermiCG.expand_each_fock_space!

FermiCG.expand_to_full_space!

FermiCG.hosvd

FermiCG.open_matvec

FermiCG.open_matvec_parallel

FermiCG.open_matvec_parallel2

FermiCG.open_matvec_thread

FermiCG.project_out!

FermiCG.tpsci_ci

FermiCG.upper_bound

FermiCG.upper_bound

FermiCG.upper_bound

FermiCG.upper_bound

FermiCG.upper_bound2

FermiCG.upper_bound2

TPSCI

FermiCG.contract_matrix_element — Method

contract_matrix_element(   term::ClusteredTerm1B, 
                                cluster_ops::Vector{ClusterOps},
                                fock_bra, bra, fock_ket, ket)

Contraction for local (1body) terms. No contraction is needed, just a lookup from the correct operator

FermiCG.contract_matrix_element — Method

contract_matrix_element(   term::ClusteredTerm2B, 
                                cluster_ops::Vector{ClusterOps},
                                fock_bra, bra, fock_ket, ket)

FermiCG.contract_matrix_element — Method

contract_matrix_element(   term::ClusteredTerm3B, 
                                cluster_ops::Vector{ClusterOps},
                                fock_bra, bra, fock_ket, ket)

FermiCG.contract_matrix_element — Method

contract_matrix_element(   term::ClusteredTerm4B, 
                                cluster_ops::Vector{ClusterOps},
                                fock_bra, bra, fock_ket, ket)

FermiCG.contract_matvec — Method

contract_matvec(    term::ClusteredTerm2B, 
                    cluster_ops::Vector{ClusterOps},
                    fock_bra, fock_ket, ket)

FermiCG.contract_matvec — Method

contract_matvec(    term::ClusteredTerm2B, 
                    cluster_ops::Vector{ClusterOps},
                    fock_bra, fock_ket, ket)

FermiCG.contract_matvec — Method

contract_matvec(    term::ClusteredTerm4B, 
                    cluster_ops::Vector{ClusterOps},
                    fock_bra, fock_ket, ket)

This version should only use M^2N^2 storage, and n^5 scaling n={MN}

FermiCG.contract_matvec — Method

contract_matvec(    term::ClusteredTerm4B, 
                    cluster_ops::Vector{ClusterOps},
                    fock_bra, fock_ket, ket)

This version should only use M^2N^2 storage, and n^5 scaling n={MN}

FermiCG.contract_matvec_M3 — Method

contract_matvec(    term::ClusteredTerm3B, 
                    cluster_ops::Vector{ClusterOps},
                    fock_bra, fock_ket, ket)

FermiCG.contract_matvec_M4 — Method

contract_matvec(    term::ClusteredTerm4B, 
                    cluster_ops::Vector{ClusterOps},
                    fock_bra, fock_ket, ket)

FermiCG.upper_bound — Method

Return upper bound on the size of matrix elements resulting from matrix multiply

V[I,J] = g1[i,I] * g2[i,J] * c

max(|V|) <= sum_i max|g1[i,:]| * max|g2[i,:]| * |c|

FermiCG.upper_bound — Method

Return upper bound on the size of tensor elements resulting from the following contraction

V[I,J] = v[i,j] * g1[i,I] * g2[j,J]

max(|V|) <= sumij |v[ij]| * |g1[i,:]|8 * |g2[j,:]|_8 * |c|

FermiCG.upper_bound — Method

Return upper bound on the size of tensor elements resulting from the following contraction

V[I,J,K] = v[i,j,k] * g1[i,I] * g2[j,J] * g3[k,K]

max(|V|) <= sumijk |v[ijk]| * |g1[i,:]|8 * |g2[j,:]|8 * |g3[k,:]|8 * |c|

FermiCG.upper_bound — Method

Return upper bound on the size of tensor elements resulting from the following contraction

V[I,J,K,L] = v[i,j,k,l] * g1[i,I] * g2[j,J] * g3[k,K] * g4[l,L]

max(|V|) <= sumijkl |v[ijkl]| * |g1[i,:]|8 * |g2[j,:]|8 * |g3[k,:]|8 * |g4[l,:]|_8

FermiCG.upper_bound2 — Method

max(HIJ(K)|K <= sumr (sumpq vpqrs max(g1[p,:]) * max(g2[q,:]) * |c| ) * |g3(r,K)|

FermiCG.upper_bound2 — Method

max(HIJK(L)|L <= sums (sumpqr vpqrs max(g1[p,:]) * max(g2[q,:]) * max(g3[r,:]) * |c| ) * |g4(s,L)|

FermiCG.build_brdm — Method

build_brdm(ci_vector::ClusteredState, ci, dims)

Build block reduced density matrix for Cluster, ci

ci_vector::ClusteredState = input state

ci = Cluster type for whihch we want the BRDM

dims = list of dimensions for each fock sector

FermiCG.build_full_H — Method

build_full_H(ci_vector::ClusteredState, cluster_ops, clustered_ham::ClusteredOperator)

Build full TPSCI Hamiltonian matrix in space spanned by ci_vector. This works in serial for the full matrix

FermiCG.build_full_H_parallel — Method

build_full_H(ci_vector::ClusteredState, cluster_ops, clustered_ham::ClusteredOperator)

Build full TPSCI Hamiltonian matrix in space spanned by ci_vector. This works in serial for the full matrix

FermiCG.compute_diagonal — Method

compute_diagonal(vector::ClusteredState{T,N,R}, cluster_ops, clustered_ham) where {T,N,R}

Form the diagonal of the hamiltonan, clustered_ham, in the basis defined by vector

FermiCG.compute_expectation_value — Method

FermiCG.compute_pt2 — Method

FermiCG.expand_each_fock_space! — Method

expand_each_fock_space!(s::ClusteredState, bases)

For each fock space sector defined, add all possible basis states

basis::Vector{ClusterBasis}

FermiCG.expand_to_full_space! — Method

expand_to_full_space(s::ClusteredState, bases)

Define all possible fock space sectors and add all possible basis states

basis::Vector{ClusterBasis}

na: Number of alpha electrons total

nb: Number of alpha electrons total

FermiCG.hosvd — Method

hosvd(ci_vector::ClusteredState, cluster_ops; hshift=1e-8, truncate=-1)

Peform HOSVD aka Tucker Decomposition of ClusteredState

FermiCG.open_matvec — Method

open_matvec(ci_vector::ClusteredState, cluster_ops, clustered_ham; thresh=1e-9, nbody=4)

Compute the action of the Hamiltonian on a tpsci state vector. Open here, means that we access the full FOIS (restricted only by thresh), instead of the action of H on v within a subspace of configurations. This is essentially used for computing a PT correction outside of the subspace, or used for searching in TPSCI.

FermiCG.open_matvec_parallel — Method

open_matvec_parallel(ci_vector::ClusteredState, cluster_ops, clustered_ham; thresh=1e-9, nbody=4)

Compute the action of the Hamiltonian on a tpsci state vector. Open here, means that we access the full FOIS (restricted only by thresh), instead of the action of H on v within a subspace of configurations. This is essentially used for computing a PT correction outside of the subspace, or used for searching in TPSCI.

This parallellizes over FockConfigs in the output state, so it's not the most fine-grained, but it avoids data races in filling the final vector

FermiCG.open_matvec_parallel2 — Method

FermiCG.open_matvec_thread — Method

open_matvec_thread(ci_vector::ClusteredState, cluster_ops, clustered_ham; thresh=1e-9, nbody=4)

Compute the action of the Hamiltonian on a tpsci state vector. Open here, means that we access the full FOIS (restricted only by thresh), instead of the action of H on v within a subspace of configurations. This is essentially used for computing a PT correction outside of the subspace, or used for searching in TPSCI.

This parallellizes over FockConfigs in the output state, so it's not the most fine-grained, but it avoids data races in filling the final vector

FermiCG.project_out! — Method

project_out!(v::ClusteredState, w::ClusteredState)

Project w out of v |v'> = |v> - |w><w|v>

FermiCG.tpsci_ci — Method

tpsci_ci(ci_vector::ClusteredState{T,N,R}, cluster_ops, clustered_ham::ClusteredOperator;
            thresh_cipsi = 1e-2,
            thresh_foi   = 1e-6,
            thresh_asci  = 1e-2,
            max_iter     = 10,
            conv_thresh  = 1e-4) where {T,N,R}

Run TPSCI